Refactor into traits
