+++ src/Factories/Concerns/Publishable.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Aerni\Factory\Factories\Concerns;
+
+trait Publishable
+{
+    public function unpublished(): self
+    {
+        return $this->set('published', false);
+    }
+}
